requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
libsql-client
torch
torchvision
opencv-python
pillow
numpy